server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.10.11",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "azure-communication-callautomation==1.5.0",
    "azure-communication-identity==1.4.0",
    "azure-core==1.37.0",
    "azure-identity==1.19.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cryptography==46.0.3",
    "fastapi==0.115.6",
    "h11==0.16.0",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "isodate==0.7.2",
    "msal==1.34.0",
    "msal-extensions==1.3.1",
    "msrest==0.7.1",
    "numpy==2.4.0",
    "oauthlib==3.3.1",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyjwt==2.10.1",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "soxr==1.0.0",
    "starlette==0.41.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uvicorn[standard]==0.32.1",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==14.1",
]
